MFC r243322 (by marcel):
Unbreak amd64 cross-build where amd64 is the target. While clang
may be installed as cc and we don't need to build gcc as a
cross-tools, we still build gcc and thus need cc_tools built
as a build tool. Not doing this results in building gengenrtl with
the target compiler while we need to run it on the build machine.


git-svn-id: https://svn.freebsd.org/base/stable/9@263751 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f