MFC r263774:
Revert r263694, and apply a better fix to squelch unnecessary warnings
from clang about possible keywords being treated as identifiers for the
remainder of the translation unit (a.k.a. -Wkeyword-compat), when using
libstdc++ in combination with -Wsystem-headers.  This will not only fix
devd, but any C++ program using libstdc++.


git-svn-id: https://svn.freebsd.org/base/stable/10@263911 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f