Correct the actual definition of ath_tx_tid_filt_comp_single() to match how it's used.
This is another bug that led to aggregate traffic hanging because
the BAW tracking stopped being accurate.  In this instance, a filtered
frame that exceeded retries would return a non-error, which would
mean the caller would never remove it from the BAW.  But it wouldn't
be added to the filtered list, so it would be lost forever.  There'd
thus be a hole in the BAW that would never get transmitted and
this leads to a traffic hang.

Tested:

* Routerstation Pro, AR9220 AP


git-svn-id: https://svn.freebsd.org/base/head@264254 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f