Add Stacey Son's binary activation patches that allow remapping of execution to a emumation program via parsing of ELF header information.
With this kernel module and userland tool, poudriere is able to build
ports packages via the QEMU userland tools (or another emulator program)
in a different architecture chroot, e.g. TARGET=mips TARGET_ARCH=mips

I'm not connecting this to GENERIC for obvious reasons, but this should
allow the kernel module to be built by default and enable the building
of the userland tool (which automatically loads the kernel module).

Submitted by:	sson@
Reviewed by:	jhb@


git-svn-id: https://svn.freebsd.org/base/head@264269 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f