login: Remove broken dialup log message.
For 10 years, the "DIALUP <tty>, <user>" message has required having a
hostname (-h) instead of not having a hostname; therefore, it is never
logged. Given that dialup is obsolete and this has not been fixed, remove
the log message.

Note that LOGALL, which is defined by default, logs a message for all
logins, including dialup logins.


git-svn-id: https://svn.freebsd.org/base/head@264677 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f