MFC r244015 (by ken): Fix the CTL OOA queue dumping code so that it does not hold a mutex while doing a copyout.  That can cause a panic, because copyout can trigger VM faults, and we can't handle VM faults while holding a mutex.
The solution here is to malloc a separate buffer to hold the OOA
queue entries, so that we don't risk a VM fault while filling up
the buffer and we don't have to drop the lock.  The other solution
would be to wire the user's memory while filling their buffer with
copyout, but that would have been a little more complex.

Also fix a debugging parenthesis issue in ctl_abort_task() pointed
out by Chuck Tuffli.


git-svn-id: https://svn.freebsd.org/base/stable/9@265645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f